refactor(settings): reorganize tab structure and naming
- Rename SettingsTab → IntegrationsTab, AppSettingsTab → SettingsTab
- Tab values: issue-settings → integrations, app-settings → settings
- Tab labels: 연동 설정 → 연동, 앱 설정 → 설정
- Merge platform subsections (project/repo/team + issue settings → single settings section)
- Move issue composition + title prefix to Settings (app-level)
